server/requirements.txt
africastalking==2.0.2
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography @ file:///home/builder/.termux-build/python-cryptography/src
dj-rest-auth==7.1.1
Django==6.0.3
django-allauth==65.15.0
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-filter==25.2
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
fido2==2.1.1
idna==3.11
inflection==0.5.1
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
pyotp==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==8.2
redis==7.3.0
requests==2.32.5
responses==0.26.0
schema==0.7.8
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
